python/pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "pptx-dev"
version = "0.1.0"
description = "Official Python SDK for pptx.dev — generate, parse, validate, convert, and render PowerPoint presentations via the OPF (Open Presentation Format)."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "pptx.dev" }]
keywords = [
  "pptx",
  "powerpoint",
  "opf",
  "presentation",
  "slides",
  "ai",
  "agents",
  "mcp",
  "langchain",
  "llamaindex",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Office/Business",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Typing :: Typed",
]
dependencies = [
  "httpx>=0.27,<1.0",
  "pydantic>=2.6,<3.0",
  "typing-extensions>=4.10; python_version<'3.11'",
]

[project.optional-dependencies]
langchain = ["langchain-core>=0.2"]
llamaindex = ["llama-index-core>=0.10"]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "respx>=0.21",
  "mypy>=1.10",
  "ruff>=0.5",
  "build>=1.2",
  "twine>=5.1",
]

[project.urls]
Homepage = "https://www.pptx.dev"
Documentation = "https://www.pptx.dev/docs/sdk/python"
Repository = "https://github.com/Data-Advantage/pptx-dev"
Issues = "https://github.com/Data-Advantage/pptx-dev/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/pptx_dev"]

[tool.hatch.build.targets.sdist]
include = ["src/pptx_dev", "README.md", "LICENSE", "pyproject.toml"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "W"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_ignores = true
